requirements.txt
aiogram            # Telegram bot framework
asyncpg           # Asynchronous PostgreSQL driver
pgvector          # Integration with PGVector for vector search
openai            # OpenAI API client for embedding generation
python-dotenv      # Load environment variables from .env file